+++ pkg/providers/ingress/translation/annotations/cors.go
